Create new abstract metadata table from sync rules
